fix: fix deploy and ci scripts
